Add integer to numeral conversion
Previously if the user searched for e.g. 'assassins creed 2', no
results would be returned. A check has been added to the search
function in the event of no match to see if the search terms
include an integer. If they do, the search is re-run with the
integer converted to roman numeral.

This will improve end user experience as many users are likely
to search using integers as terms.