Add animation, styling to snipe results
Added additional class to distinguish sub-results from
headline game results in the snipe route.

This makes it clearer that when the user clicks on 'expand',
they are viewing all sales logged for that game. Previously
the divs were styled identically.

Also added animation function using ajax. This triggers
when the user clicks 'expand'. The opacity gradually
increases, again making it clearer for the user that 
they are viewing sub-results.